FinalRag/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-huggingface
langchain-ollama
langchain-openai
chromadb
spacy
python-dotenv
pydantic
supabase
pdfplumber
langchain_chroma
sentence-transformers
openai